This is a slightly opinionated commit, consisting of three main changes:
First, superficial things. Model is renamed to Table, and there's no Model alias (why have a multiple names for things when it isn't necessary?). It also renamed all to select (which is what the test calls it anyways).

Seconly, it attempts to make the API more unified for count, select and destroy. These take an optional single where parameter. To specify explicit columns in select, use the colums method of query.

Finally, the above has the advantage of making state within Table, via the now deceased parseArgs, a thing of the past. A table is now meant to be a singletone (next branch will explore a way to do this).

I plan to revisit [batch]Insert and update with the same enthusiasm at some point in the near future.