README: fix `deasync` repo link and correct a few typos
The link to `deasync` does not point to the most up-to-date version of
the package.  It appears that @vkurchatkin has stopped maintaining it
and given ownership to @abbr instead.  This caused me to unnecessarily
open vkurchatkin/deasync#1.

This patch corrects the link, which should prevent any future confusion
and also fixes a few typos and markdown formatting issues.